bue merdas

+++ src/Game/Poligono.java
@@ -236,7 +236,7 @@ public class Poligono {
      * @param y Quantidade a ser transladada na direção y.
      * @return Novo polígono resultante da translação.
      */
-    public Poligono translacao(int x, int y)
+    public Poligono translacao(double x, double y)
     {
         ArrayList<Ponto> newPontos = new ArrayList<>();
         for (Ponto ponto : pontos) {
@@ -252,12 +252,12 @@ public class Poligono {
      * @param novoCentroideY Nova coordenada y para o centroide do polígono.
      * @return Novo polígono resultante da translação do centroide para a nova posição.
      */
-    public Poligono translacaoCentroide(int novoCentroideX, int novoCentroideY){
+    public Poligono translacaoCentroide(double novoCentroideX, double novoCentroideY){
 
         ArrayList<Ponto> newPontos = new ArrayList<>();
         Ponto centroidePoligono = calcularCentro();
-        int deslocamentoX = novoCentroideX - centroidePoligono.getX();
-        int deslocamentoY = novoCentroideY - centroidePoligono.getY();
+        double deslocamentoX = novoCentroideX - centroidePoligono.getX();
+        double deslocamentoY = novoCentroideY - centroidePoligono.getY();
 
         return translacao(deslocamentoX, deslocamentoY);
