+++ src/Game/Retangulo.java
@@ -134,7 +134,7 @@ public class Retangulo extends Poligono {
      * @return Novo retângulo resultante da translação.
      */
     @Override
-    public Retangulo translacao(int x, int y)
+    public Retangulo translacao(double x, double y)
     {
         ArrayList<Ponto> newPontos = new ArrayList<>();
         for (Ponto ponto : pontos) {
@@ -151,12 +151,12 @@ public class Retangulo extends Poligono {
      * @return Novo retângulo resultante da translação do centroide para a nova posição.
      */
     @Override
-    public Retangulo translacaoCentroide(int novoCentroideX, int novoCentroideY){
+    public Retangulo translacaoCentroide(double novoCentroideX, double novoCentroideY){
 
         ArrayList<Ponto> newPontos = new ArrayList<>();
         Ponto centroidePoligono = calcularCentro();
-        int deslocamentoX = novoCentroideX - centroidePoligono.getX();
-        int deslocamentoY = novoCentroideY - centroidePoligono.getY();
+        double deslocamentoX = novoCentroideX - centroidePoligono.getX();
+        double deslocamentoY = novoCentroideY - centroidePoligono.getY();
 
         return translacao(deslocamentoX, deslocamentoY);
